WIP: for full update script implementation
Verify and updates the platform-tools to the latest version if outdated/not already installed.

Verify and reboot to fastboot if the user is not already in fastboot mode.

Verify the user has the latest bootloader/radio and update to the lastest version packed within the zip if the version they have is outdated.

Finally, verify if the user wants to wipe/not wipe and updates.

Change-Id: I5e279ab6de3df4ffff640f5ebfcde067b98b36b1
Signed-off-by: Alvin Francis <nivlafx@gmail.com>